if file is already saved, skips the downloading

+++ comics.py
@@ -5,6 +5,14 @@ import os
 import urllib.request
 import webbrowser
 
+
+def fileExistence(title):
+    for x in os.listdir('./pics/'):
+        if x == (title + '.png'):
+            return True
+    return False
+
+
 startInt = random.randint(1, 2187)
 
 print('How many comics do you want(1-100)')
@@ -24,7 +32,7 @@ for i in range(startInt, startInt + int(numCom)):
     nameOfFile = elem[0].text.replace(' ', '_').replace('\'', '').lower()
     webbrowser.open('https://imgs.xkcd.com/comics/' + nameOfFile + '.png')
     print(nameOfFile)
-    if save.lower() == 'y':
+    if save.lower() == 'y' and not fileExistence(nameOfFile):
         urllib.request.urlretrieve(
             'https://imgs.xkcd.com/comics/' + nameOfFile + '.png', './pics/' + nameOfFile + '.png')
         print(nameOfFile + ' comic saved!')
